+++ CHANGELOG.md
@@ -9,9 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - **[BREAKING]** Drop Python 3.8 support. Minimum Python version is 3.9 now.
-
-Nothing unreleased!
-
+- changing the default host from 0.0.0.0 to 127.0.0.1  
+  
 ## [1.1.403rc0] - 2024-08-13
 
 ### Fixed